requirements.txt
streamlit
pandas
twilio
flask
biopython
pillow
piexif
